extract filters to components
